+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-10 20:36:10
-// Commit: test(state): add unit tests for state business logic
+// Updated at: 2026-05-10 22:00:35
+// Commit: feat(state): integrate custom telemetry logger in state
 
 class stateComponentState {
-  final String commitRef = '1035';
-  final String lastMessage = 'test(state): add unit tests for state business logic';
-  final DateTime timestamp = DateTime.parse('2026-05-10T20:36:10+05:00');
+  final String commitRef = '1041';
+  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
+  final DateTime timestamp = DateTime.parse('2026-05-10T22:00:35+05:00');
 }